requirements.txt
aiohttp
attrs
mypy
orjson
